Adding tiny card and manage button
